Update frontend gql types

+++ apps/ui/modules/queries/reservationUnit.tsx
@@ -196,7 +196,7 @@ export const SEARCH_RESERVATION_UNITS = gql`
     $textSearch: String
     $pk: [Int]
     $applicationRound: [Int]
-    $personsAllowed: Decimal
+    $personsAllowed: Int
     $unit: [Int]
     $reservationUnitType: [Int]
     $purposes: [Int]
@@ -208,7 +208,7 @@ export const SEARCH_RESERVATION_UNITS = gql`
     $reservableDateEnd: Date
     $reservableTimeStart: Time
     $reservableTimeEnd: Time
-    $reservableMinimumDurationMinutes: Decimal
+    $reservableMinimumDurationMinutes: Int
     $showOnlyReservable: Boolean
     $first: Int
     $before: String
